# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Agent Tool Chat API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /mcp-servers:
    get:
      operationId: listMcpServers
      tags: [mcpServers]
      summary: List all MCP servers
      responses:
        "200":
          description: List of MCP servers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/McpServer"
    post:
      operationId: createMcpServer
      tags: [mcpServers]
      summary: Create a new MCP server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMcpServerBody"
      responses:
        "201":
          description: Created MCP server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McpServer"
  /mcp-servers/{id}:
    get:
      operationId: getMcpServer
      tags: [mcpServers]
      summary: Get a specific MCP server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: MCP server details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McpServer"
        "404":
          description: Not found
    patch:
      operationId: updateMcpServer
      tags: [mcpServers]
      summary: Update an MCP server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMcpServerBody"
      responses:
        "200":
          description: Updated MCP server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McpServer"
    delete:
      operationId: deleteMcpServer
      tags: [mcpServers]
      summary: Delete an MCP server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /mcp-servers/{id}/test:
    post:
      operationId: testMcpServerConnection
      tags: [mcpServers]
      summary: Test MCP server connection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Connection test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McpConnectionTestResult"
  /mcp-servers/{id}/discover:
    post:
      operationId: discoverMcpTools
      tags: [mcpServers]
      summary: Discover tools from an MCP server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Discovered tools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/McpTool"
  /mcp-servers/{id}/tools:
    get:
      operationId: listMcpTools
      tags: [mcpTools]
      summary: List tools for an MCP server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of tools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/McpTool"
  /mcp-servers/{id}/prompts:
    get:
      operationId: listMcpPrompts
      tags: [mcpTools]
      summary: List prompts discovered from an MCP server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of prompts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/McpPrompt"
  /mcp-servers/{id}/resources:
    get:
      operationId: listMcpResources
      tags: [mcpTools]
      summary: List resources discovered from an MCP server
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of resources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/McpResource"
  /mcp-tools/{toolId}:
    patch:
      operationId: updateMcpTool
      tags: [mcpTools]
      summary: Update an MCP tool (enable/disable, require approval)
      parameters:
        - name: toolId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMcpToolBody"
      responses:
        "200":
          description: Updated tool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McpTool"
...
components:
  schemas:
...
    McpServer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        transportType:
          type: string
          enum: [stdio, streamable-http]
        endpoint:
          type: string
        command:
          type: string
        args:
          type: array
          items:
            type: string
        authType:
          type: string
          enum: [none, bearer, api-key]
        timeout:
          type: integer
        retryCount:
          type: integer
        status:
          type: string
          enum: [connected, disconnected, checking, error]
        enabled:
          type: boolean
        toolCount:
          type: integer
        lastCheckedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - transportType
        - status
        - enabled
        - toolCount
        - createdAt
    CreateMcpServerBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        transportType:
          type: string
          enum: [stdio, streamable-http]
        endpoint:
          type: string
        command:
          type: string
        args:
          type: array
          items:
            type: string
        authType:
          type: string
          enum: [none, bearer, api-key]
        authSecret:
          type: string
        timeout:
          type: integer
          default: 30
        retryCount:
          type: integer
          default: 3
        enabled:
          type: boolean
          default: true
      required:
        - name
        - transportType
    UpdateMcpServerBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        transportType:
          type: string
          enum: [stdio, streamable-http]
        endpoint:
          type: string
        command:
          type: string
        args:
          type: array
          items:
            type: string
        authType:
          type: string
          enum: [none, bearer, api-key]
        authSecret:
          type: string
        timeout:
          type: integer
        retryCount:
          type: integer
        enabled:
          type: boolean
    McpConnectionTestResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        toolCount:
          type: integer
        latencyMs:
          type: integer
      required:
        - success
        - message
    McpTool:
      type: object
      properties:
        id:
          type: integer
        serverId:
          type: integer
        serverName:
          type: string
        toolName:
          type: string
        description:
          type: string
        inputSchema:
          type: object
        outputSchema:
          type: object
        enabled:
          type: boolean
        requiresApproval:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - serverId
        - toolName
        - enabled
        - requiresApproval
        - createdAt
    McpResource:
      type: object
      properties:
        id:
          type: integer
        serverId:
          type: integer
        resourceName:
          type: string
        description:
          type: string
        resourceType:
          type: string
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - serverId
        - resourceName
        - resourceType
        - createdAt
    McpPrompt:
      type: object
      properties:
        id:
          type: integer
        serverId:
          type: integer
        promptName:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - serverId
        - promptName
        - createdAt
    UpdateMcpToolBody:
      type: object
      properties:
        enabled:
          type: boolean
        requiresApproval:
          type: boolean
...