# lib/api-zod/src/generated/types/mcpPrompt.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent Tool Chat API specification
 * OpenAPI spec version: 0.1.0
 */

export interface McpPrompt {
  id: number;
  serverId: number;
  promptName: string;
  description?: string;
  createdAt: Date;
}