# lib/api-zod/src/generated/types/mcpServer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Agent Tool Chat API specification
 * OpenAPI spec version: 0.1.0
 */
import type { McpServerAuthType } from "./mcpServerAuthType";
import type { McpServerStatus } from "./mcpServerStatus";
import type { McpServerTransportType } from "./mcpServerTransportType";

export interface McpServer {
  id: number;
  name: string;
  description?: string;
  transportType: McpServerTransportType;
  endpoint?: string;
  command?: string;
  args?: string[];
  authType?: McpServerAuthType;
  timeout?: number;
  retryCount?: number;
  status: McpServerStatus;
  enabled: boolean;
  toolCount: number;
  lastCheckedAt?: Date;
  createdAt: Date;
}